refactor: move sqllog.go to geopackage since we don't use it in other data sources
