+++ internal/ogc/features/datasources/geopackage/geopackage.go
@@ -40,7 +40,7 @@ func loadDriver() {
 	once.Do(func() {
 		spatialite := path.Join(os.Getenv("SPATIALITE_LIBRARY_PATH"), "mod_spatialite")
 		driver := &sqlite3.SQLiteDriver{Extensions: []string{spatialite}}
-		sql.Register(sqliteDriverName, sqlhooks.Wrap(driver, ds.NewSQLLogFromEnv()))
+		sql.Register(sqliteDriverName, sqlhooks.Wrap(driver, NewSQLLogFromEnv()))
 	})
 }
